fix: Ботинки клоуна сново работают
