+++ Content.Server/_White/Animations/WaddleAnimationSystem.cs
@@ -1,0 +1,16 @@
+﻿using Content.Shared._White.Animations;
+
+namespace Content.Server._White.Animations;
+
+public sealed class WaddleAnimationSystem : SharedWaddledAnimationSystem
+{
+    protected override void PlayAnimation(EntityUid user)
+    {
+        RaiseNetworkEvent(new StartedWaddlingEvent(GetNetEntity(user)));
+    }
+
+    protected override void StopAnimation(EntityUid user)
+    {
+        RaiseNetworkEvent(new StoppedWaddlingEvent(GetNetEntity(user)));
+    }
+}
