+++ Content.Shared/Clothing/EntitySystems/WaddleClothingSystem.cs
@@ -1,8 +1,8 @@
-﻿using Content.Shared.Clothing.Components;
-using Content.Shared.Movement.Components;
+﻿using Content.Shared._White.Animations;
+using Content.Shared.Clothing.Components;
 using Content.Shared.Inventory.Events;
 
-namespace Content.Client.Clothing.Systems;
+namespace Content.Shared.Clothing.EntitySystems;
 
 public sealed class WaddleClothingSystem : EntitySystem
 {
